requirements/common.txt
django-extensions
django-debug-toolbar
django-bootstrap-v5~=-1.0.11
django-pydenticon~=0.2
jupyter
pillow
easy-thumbnails~=2.8.3

